refactor: Move create_appmap_file and its dependencies from testing_framework to web_framework
