+++ appmap/_implementation/web_framework.py
@@ -1,11 +1,18 @@
 """Common utilities for web framework integration"""
 
+import datetime
+import os
+import os.path
 import re
 import time
+from hashlib import sha256
+from tempfile import NamedTemporaryFile
 
+from appmap._implementation import generation
+from appmap._implementation.env import Env
 from appmap._implementation.event import Event, ReturnEvent, describe_value
 from appmap._implementation.recording import Recorder
-from appmap._implementation.utils import root_relative_path
+from appmap._implementation.utils import root_relative_path, scenario_filename
 
 
 class TemplateEvent(Event):  # pylint: disable=too-few-public-methods
@@ -56,3 +63,61 @@ class TemplateHandler:  # pylint: disable=too-few-public-methods
         finally:
             if rec.enabled:
                 Recorder.add_event(ReturnEvent(call_event.id, time.monotonic() - start))
+
+
+NAME_MAX = 255  # true for most filesystems
+HASH_LEN = 7  # arbitrary, but git proves it's a reasonable value
+APPMAP_SUFFIX = ".appmap.json"
+
+
+def name_hash(namepart):
+    """Returns the hex digits of the sha256 of the os.fsencode()d namepart."""
+    return sha256(os.fsencode(namepart)).hexdigest()
+
+
+def write_appmap(basedir, basename, contents):
+    """Write an appmap file into basedir.
+
+    Adds APPMAP_SUFFIX to basename; shortens the name if necessary.
+    Atomically replaces existing files. Creates the basedir if required.
+    """
+
+    if len(basename) > NAME_MAX - len(APPMAP_SUFFIX):
+        part = NAME_MAX - len(APPMAP_SUFFIX) - 1 - HASH_LEN
+        basename = basename[:part] + "-" + name_hash(basename[part:])[:HASH_LEN]
+    filename = basename + APPMAP_SUFFIX
+
+    if not basedir.exists():
+        basedir.mkdir(parents=True, exist_ok=True)
+
+    with NamedTemporaryFile(mode="w", dir=basedir, delete=False) as tmp:
+        tmp.write(contents)
+    os.replace(tmp.name, basedir / filename)
+
+
+def create_appmap_file(
+    request_method, request_path_info, request_full_path, response, headers, rec
+):
+    start_time = datetime.datetime.now()
+    appmap_name = (
+        request_method
+        + " "
+        + request_path_info
+        + " ("
+        + str(response.status_code)
+        + ") - "
+        + start_time.strftime("%T.%f")[:-3]
+    )
+    output_dir = Env.current.output_dir
+    appmap_basename = scenario_filename(
+        "_".join([str(start_time.timestamp()), request_full_path])
+    )
+    appmap_file_path = os.path.join(output_dir, appmap_basename)
+    metadata = {
+        "name": appmap_name,
+        "timestamp": start_time.timestamp(),
+        "recorder": {"name": "record_requests"},
+    }
+    write_appmap(output_dir, appmap_basename, generation.dump(rec, metadata))
+    headers["AppMap-Name"] = os.path.abspath(appmap_name)
+    headers["AppMap-File-Name"] = os.path.abspath(appmap_file_path) + APPMAP_SUFFIX
